@@ -1,13 +1,13 @@
 import UIKit
 import Flutter
-
+import GoogleMaps
 @UIApplicationMain
 @objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
-     GMSServices.provideAPIKey("AIzaSyBago92YWfXzXsILKI_7D2yGw-SH54wuTQ")
+     GMSServices.provideAPIKey("AIzaSyCkGTIrK_ko923z2nIp8Cica9sEw2Z1tvw")
     GeneratedPluginRegistrant.register(with: self)
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
   }
