@@ -1,9 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:khosousi_online/core/managers/color_manager.dart';
+import 'package:khosousi_online/core/managers/string_manager.dart';
+import 'package:khosousi_online/core/ui/dialogs_widgets/loading_content_dialog.dart';
 import 'package:khosousi_online/core/ui/style/common_styles.dart';
 import 'package:khosousi_online/core/ui/widgets/custom_elevated_btn.dart';
 import 'package:khosousi_online/core/ui/widgets/custom_text_field.dart';
+import 'package:khosousi_online/core/locator/service_locator.dart' as sl;
+import 'package:khosousi_online/core/utils/helpers/snackbar.dart';
+import 'package:khosousi_online/core/utils/helpers/toast_utils.dart';
+import 'package:khosousi_online/features/accounts/domain/repositories/auth_repo.dart';
+import 'package:khosousi_online/features/edit_password/presentation/cubit/change_password_cubit.dart';
 
 class EditPasswordScreen extends StatelessWidget {
   static const routeName = 'edit_password_screen';
@@ -11,44 +19,79 @@ class EditPasswordScreen extends StatelessWidget {
   final _formKey = GlobalKey<FormState>();
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: Text('تغيير كلمة المرور'),
-        ),
-        body: SafeArea(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: SingleChildScrollView(
-              child: Column(
-                children: [
-                  Container(
-                    child: Form(
-                      key: _formKey,
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          _buildCurrentPasswordTextField(context),
-                          SizedBox(
-                            height: 16.h,
+    return BlocProvider(
+      create: (context) => sl.locator<ChangePasswordCubit>(),
+      child: Builder(builder: (context) {
+        return BlocListener<ChangePasswordCubit, ChangePasswordState>(
+          listener: (context, state) {
+            _buildListener(state, context);
+          },
+          child: Scaffold(
+            backgroundColor: Colors.white,
+              appBar: AppBar(
+                title: Text('تغيير كلمة المرور'),
+              ),
+              body: SafeArea(
+                child: Padding(
+                  padding: const EdgeInsets.all(16.0),
+                  child: SingleChildScrollView(
+                    child: Column(
+                      children: [
+                        Container(
+                          child: Form(
+                            key: _formKey,
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.stretch,
+                              children: [
+                                _buildCurrentPasswordTextField(context),
+                                SizedBox(
+                                  height: 16.h,
+                                ),
+                                _buildNewPasswordTextField(context),
+                                SizedBox(
+                                  height: 16.h,
+                                ),
+                                _buildConfirmPasswordTextField(context),
+                                SizedBox(
+                                  height: 24.h,
+                                ),
+                                _buildBtns(context)
+                              ],
+                            ),
                           ),
-                          _buildNewPasswordTextField(context),
-                          SizedBox(
-                            height: 16.h,
-                          ),
-                          _buildConfirmPasswordTextField(context),
-                          SizedBox(
-                            height: 24.h,
-                          ),
-                          _buildBtns(context)
-                        ],
-                      ),
+                        ),
+                      ],
                     ),
                   ),
-                ],
-              ),
-            ),
-          ),
-        ));
+                ),
+              )),
+        );
+      }),
+    );
+  }
+
+  void _buildListener(ChangePasswordState state, BuildContext context) {
+    if (state.changePasswordStatus == ChangePasswordStatus.loading) {
+      showDialog(
+        context: context,
+        barrierDismissible: false,
+        builder: (context) {
+          return Dialog(
+              child: LoadingContentDialog(text: 'جاري تغيير كلمة المرور'));
+        },
+      );
+    } else if (state.changePasswordStatus == ChangePasswordStatus.noInternet) {
+      Navigator.pop(context);
+      showSnackbar(context, AppStrings.noInternetConnectionMessage);
+    } else if (state.changePasswordStatus ==
+        ChangePasswordStatus.networkError) {
+      Navigator.pop(context);
+      showSnackbar(context, state.errorMessage);
+    } else if (state.changePasswordStatus == ChangePasswordStatus.done) {
+      Navigator.pop(context);
+      ToastUtils.showSusToastMessage('تم تغيير كلمة المرور بنجلح');
+      Navigator.pop(context);
+    }
   }
 
   Row _buildBtns(BuildContext context) {
@@ -59,7 +102,17 @@ class EditPasswordScreen extends StatelessWidget {
         CustomElevatedButton(
           backgroundColor: ColorManager.black,
           label: 'تغيير كلمة المرور',
-          onPressed: () {},
+          onPressed: () {
+            if (_formKey.currentState!.validate()) {
+              if (context.read<ChangePasswordCubit>().state.newPassword !=
+                  context.read<ChangePasswordCubit>().state.confirmPassword) {
+                showSnackbar(context, 'تأكد بأن كلمتي المرور متطابقتان');
+              } else {
+                BlocProvider.of<ChangePasswordCubit>(context)
+                    .submit(context.read<AuthRepo>().getUserId()!);
+              }
+            }
+          },
         )
       ],
     );
@@ -80,11 +133,19 @@ class EditPasswordScreen extends StatelessWidget {
           textInputAction: TextInputAction.done,
           textInputType: TextInputType.text,
           hintText: '********',
-          validator: (value) {},
+          validator: (value) {
+            if (value == null || value.isEmpty) {
+              return 'عليك إدخال هذا الحقل';
+            }
+          },
           isObscure: true,
           iconData: Icon(Icons.lock),
           withBorder: true,
           isFilled: true,
+          onChanged: (value) {
+            BlocProvider.of<ChangePasswordCubit>(context)
+                .changeOldPassword(value!);
+          },
         ),
       ],
     );
@@ -105,11 +166,19 @@ class EditPasswordScreen extends StatelessWidget {
           textInputAction: TextInputAction.done,
           textInputType: TextInputType.text,
           hintText: '********',
-          validator: (value) {},
+          validator: (value) {
+            if (value == null || value.isEmpty) {
+              return 'عليك إدخال هذا الحقل';
+            }
+          },
           isObscure: true,
           iconData: Icon(Icons.lock),
           withBorder: true,
           isFilled: true,
+          onChanged: (value) {
+            BlocProvider.of<ChangePasswordCubit>(context)
+                .changeNewPassword(value!);
+          },
         ),
       ],
     );
@@ -130,11 +199,19 @@ class EditPasswordScreen extends StatelessWidget {
           textInputAction: TextInputAction.done,
           textInputType: TextInputType.text,
           hintText: '********',
-          validator: (value) {},
+          validator: (value) {
+            if (value == null || value.isEmpty) {
+              return 'عليك إدخال هذا الحقل';
+            }
+          },
           isObscure: true,
           iconData: Icon(Icons.lock),
           withBorder: true,
           isFilled: true,
+          onChanged: (value) {
+            BlocProvider.of<ChangePasswordCubit>(context)
+                .changeConfirmNewPassword(value!);
+          },
         ),
       ],
     );
