@@ -3,14 +3,13 @@ import 'dart:convert';
 
 import 'package:equatable/equatable.dart';
 
-// "Id": "1",
+//         "Id": "1",
 //         "country_code": "AF",
 //         "country_name_en": "Afghanistan",
 //         "country_name_en_url": "Afghanistan",
 //         "country_name_ar": "",
 //         "calling_code": "93"
 class CountryEntity extends Equatable {
-
   final String id;
   final String countryCode;
   final String countryNameEn;
@@ -25,17 +24,23 @@ class CountryEntity extends Equatable {
     required this.countryNameAr,
     required this.callingCode,
   });
+  factory CountryEntity.empty() {
+    return CountryEntity(
+        callingCode: '',
+        countryCode: '',
+        countryNameAr: '',
+        countryNameEn: '',
+        country_name_en_url: '',
+        id: '');
+  }
 
   @override
-  List<Object?> get props =>
-  [
-    id,
-    countryCode,
-    countryNameEn,
-    country_name_en_url,
-    countryNameAr,
-    callingCode,
-  ];
-
-
+  List<Object?> get props => [
+        id,
+        countryCode,
+        countryNameEn,
+        country_name_en_url,
+        countryNameAr,
+        callingCode,
+      ];
 }
